Update variables.rst and operators.rst
In this commit, I have finished the section "Variables Introduction" and have started the section called "Operators and Expressions".  "Variables Introduction" introduces primitive types and Strings, as well as naming conventions for variables and Shadow reserved words. "Operators and Expressions" starts by outlining variable assignment and unique Shadow features, and then beings to explain the different types of of operators.